+++ internal/provider/filemanager_directory_resource.go
@@ -5,15 +5,13 @@ import (
 	"context"
 	"fmt"
 	"github.com/hashicorp/terraform-plugin-framework/path"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"path/filepath"
-	"strings"
-
-	"github.com/hashicorp/terraform-plugin-framework/resource"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
 )
 
 // Ensure the implementation satisfies the expected interfaces.
@@ -54,29 +52,17 @@ func (r *filemanagerDirectoryResource) Schema(_ context.Context, _ resource.Sche
 				Description: "Placeholder ID value",
 			},
 			"path": schema.StringAttribute{
-				Computed:            true,
-				MarkdownDescription: "The full path of the directory. Equal to `{parent}/{name}/`.",
-			},
-			"parent": schema.StringAttribute{
-				Required:      true,
-				PlanModifiers: []planmodifier.String{stringplanmodifier.RequiresReplace()},
-				Description: "The parent of the directory. Must include a trailing slash. " +
-					"The parent will be created as an unmanaged directory if it does not yet exist.",
-			},
-			"name": schema.StringAttribute{
-				Required:      true,
-				PlanModifiers: []planmodifier.String{stringplanmodifier.RequiresReplace()},
-				Description:   "The name of the directory.",
+				Required:            true,
+				PlanModifiers:       []planmodifier.String{stringplanmodifier.RequiresReplace()},
+				MarkdownDescription: "The full path of the directory. Must not include a trailing slash. Example: /parent/directory",
 			},
 		},
 	}
 }
 
 type filemanagerDirectoryResourceModel struct {
-	ID     types.String `tfsdk:"id"`
-	Path   types.String `tfsdk:"path"`
-	Parent types.String `tfsdk:"parent"`
-	Name   types.String `tfsdk:"name"`
+	ID   types.String `tfsdk:"id"`
+	Path types.String `tfsdk:"path"`
 }
 
 // Create creates the resource and sets the initial Terraform state.
@@ -90,10 +76,12 @@ func (r *filemanagerDirectoryResource) Create(ctx context.Context, req resource.
 	}
 
 	// Create new directory
-	parent := plan.Parent.ValueString()
-	name := plan.Name.ValueString()
-	tflog.Info(ctx, fmt.Sprintf("Creating filemanager directory %v/%v", parent, name))
-	createDirResponse, err := r.client.Files.CreateDir(parent, name)
+	directoryPath := plan.Path.ValueString()
+	pathCleaned := filepath.Clean(directoryPath)
+	pathParent := filepath.Dir(pathCleaned)
+	pathName := filepath.Base(pathCleaned)
+	tflog.Info(ctx, fmt.Sprintf("Creating filemanager directory %v/%v", pathParent, pathName))
+	_, err := r.client.Files.CreateDir(pathParent, pathName)
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Error creating filemanager_directory",
@@ -103,7 +91,6 @@ func (r *filemanagerDirectoryResource) Create(ctx context.Context, req resource.
 
 	// Map response body to schema and populate Computed attribute values
 	plan.ID = types.StringValue("placeholder")
-	plan.Path = types.StringValue(createDirResponse.Path)
 
 	// Set state to fully populated data
 	diags = resp.State.Set(ctx, plan)
@@ -124,11 +111,9 @@ func (r *filemanagerDirectoryResource) Read(ctx context.Context, req resource.Re
 	}
 
 	// Get the refreshed directory value from Qbee
-	directoryName := ""
-	directoryParent := ""
 	directoryPath := state.Path.ValueString()
 
-	listFilesResponse, err := r.client.Files.List()
+	listFilesResponse, err := r.client.Files.GetMetadata(directoryPath)
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Error reading Qbee Filemanager data",
@@ -136,20 +121,11 @@ func (r *filemanagerDirectoryResource) Read(ctx context.Context, req resource.Re
 		return
 	}
 
-	exists := false
-	for _, item := range listFilesResponse.Items {
-		if item.Path == directoryPath && item.IsDir {
-			exists = true
-			directoryName = item.Name
-			directoryParent = directoryPath[:len(directoryPath)-len(fmt.Sprintf("%v/", directoryName))]
-		}
-	}
+	exists := listFilesResponse.IsDir
 
 	// Delete from the state if it no longer exists
 	if exists {
 		state.ID = types.StringValue("placeholder")
-		state.Name = types.StringValue(directoryName)
-		state.Parent = types.StringValue(directoryParent)
 		state.Path = types.StringValue(directoryPath)
 	} else {
 		resp.State.RemoveResource(ctx)
@@ -177,7 +153,7 @@ func (r *filemanagerDirectoryResource) Delete(ctx context.Context, req resource.
 	}
 
 	// Delete old directory
-	directoryPath := strings.TrimSuffix(state.Path.ValueString(), "/")
+	directoryPath := state.Path.ValueString()
 	tflog.Info(ctx, fmt.Sprintf("Deleting filemanager directory '%v'", directoryPath))
 
 	err := r.client.Files.Delete(directoryPath)
@@ -191,12 +167,7 @@ func (r *filemanagerDirectoryResource) Delete(ctx context.Context, req resource.
 
 func (r *filemanagerDirectoryResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
 	cleanPath := filepath.Clean(req.ID)
-	filePath := cleanPath + "/"
-	fileParent := filepath.Dir(cleanPath) + "/"
-	fileName := filepath.Base(cleanPath)
 
 	resp.Diagnostics.Append(resp.State.SetAttribute(ctx, path.Root("id"), "placeholder")...)
-	resp.Diagnostics.Append(resp.State.SetAttribute(ctx, path.Root("path"), filePath)...)
-	resp.Diagnostics.Append(resp.State.SetAttribute(ctx, path.Root("parent"), fileParent)...)
-	resp.Diagnostics.Append(resp.State.SetAttribute(ctx, path.Root("name"), fileName)...)
+	resp.Diagnostics.Append(resp.State.SetAttribute(ctx, path.Root("path"), cleanPath)...)
 }